c_requirements.txt

pyaudio
paho-mqtt
aiohttp
aiortc
keyboard